Safari-609.1.11.1/LayoutTests/fast/dom/Element/id-in-frame.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<script>
description("Test using id in frame");
    
var parent = document.createElement('iframe');
parent.id = 'theframe';
document.body.appendChild(parent);

shouldBeNonNull('document.getElementById("theframe")');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
